backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.6.0
deep-sort-realtime>=1.3.0
scipy>=1.10.0
